[file-dialog] extracted LocationListRenderer
