+++ src/application/browser/widgets/index.ts
@@ -6,4 +6,5 @@
  */
 
 export * from './widget';
+export * from './virtual-renderer';
 export * from './virtual-widget';
