+++ src/application/browser/widgets/virtual-widget.ts
@@ -6,9 +6,10 @@
  */
 
 import { injectable } from "inversify";
-import { h, VirtualNode, VirtualText, VirtualDOM } from "@phosphor/virtualdom";
+import { h } from "@phosphor/virtualdom";
 import { DisposableCollection } from "../../common";
 import { BaseWidget, Message } from "./widget";
+import { VirtualRenderer } from "./virtual-renderer";
 
 @injectable()
 export class VirtualWidget extends BaseWidget {
@@ -19,7 +20,7 @@ export class VirtualWidget extends BaseWidget {
         super.onUpdateRequest(msg);
 
         const child = this.render();
-        VirtualWidget.render(child, this.node, () =>
+        VirtualRenderer.render(child, this.node, () =>
             this.onRender.dispose()
         );
     }
@@ -28,53 +29,4 @@ export class VirtualWidget extends BaseWidget {
         return null;
     }
 
-}
-
-export namespace VirtualWidget {
-    export function render(child: h.Child, host: HTMLElement, onRender?: () => void) {
-        const content = toContent(child);
-        VirtualDOM.render(content, host);
-        if (onRender) {
-            onRender();
-        }
-    }
-    export function flatten(children: h.Child[]): h.Child {
-        return children.reduce((prev, current) => merge(prev, current), null);
-    }
-
-    export function merge(left: h.Child | undefined, right: h.Child | undefined): h.Child {
-        if (!right) {
-            return left || null;
-        }
-        if (!left) {
-            return right;
-        }
-        const result = left instanceof Array ? left : [left];
-        if (right instanceof Array) {
-            result.push(...right);
-        } else {
-            result.push(right);
-        }
-        return result;
-    }
-
-    export function toContent(children: h.Child): VirtualNode | VirtualNode[] | null {
-        if (!children) {
-            return null;
-        }
-        if (typeof children === "string") {
-            return new VirtualText(children);
-        }
-        if (children instanceof Array) {
-            const nodes: VirtualNode[] = [];
-            for (const child of children) {
-                if (child) {
-                    const node = typeof child === "string" ? new VirtualText(child) : child;
-                    nodes.push(node);
-                }
-            }
-            return nodes;
-        }
-        return children;
-    }
 }
